finish the marriage before divorced judge
finish the marriage before divorced judge,add mDay,mMonth,mYear,dDay,dMonth,dYear to Family class
