+++ ParseGEDCOMFile.java
@@ -1,4 +1,3 @@
-
 import java.io.BufferedReader;
 import java.io.File;
 import java.io.FileReader;
@@ -13,13 +12,13 @@ public class ParseGEDCOMFile {
 	public static String [] level1Tag = {"NAME", "SEX", "BIRT", "DEAT", "FAMC", "FAMS", "MARR", "HUSB", "WIFE", "CHIL", "DIV"};
 	public static String [] level2Tag = {"DATE"};
 	public static String valid = "";
-	
+
 	public static void main (String [] args) {
 		HashMap<Integer,Individual> indiList = new HashMap<Integer,Individual>();
 		HashMap<Integer,Family> famList = new HashMap<Integer,Family>();
 		try {
-			File gedcom = new File ("src/input"); //must specify path			
-			BufferedReader reader = new BufferedReader(new FileReader(gedcom));		
+			File gedcom = new File ("src/input"); //must specify path
+			BufferedReader reader = new BufferedReader(new FileReader(gedcom));
 			String readLine = "";
 			Object obj = new Object();
 			int key=0;
@@ -27,7 +26,7 @@ public class ParseGEDCOMFile {
 				String splitParts [] = readLine.split(" ", 3);
 				int level;
 				String tag = "";
-				String argument = "";	
+				String argument = "";
 				level = Integer.parseInt(splitParts[0]);
 				if(level == 0){
 					if(splitParts.length > 2){
@@ -37,7 +36,7 @@ public class ParseGEDCOMFile {
 						key = Integer.parseInt(id.replaceAll("I", ""));
 						indi.setId(id);
 						indiList.put(key, indi);
-						
+
 				}else if(splitParts[2].equals("FAM")){
 						Family fam = new Family();
 						String id = splitParts[1].replaceAll("@", "");
@@ -46,8 +45,8 @@ public class ParseGEDCOMFile {
 						famList.put(key, fam);
 
 			}else{
-			}	
-			}			
+			}
+			}
 			}else if(level == 1){
 				tag = splitParts[1];
 				if(isValid(level,tag)){
@@ -68,6 +67,8 @@ public class ParseGEDCOMFile {
 						tag = splitParts[1];
 						if(level == 2 && tag.equals("DATE")){
 							argument = splitParts[2];
+							int year = Integer.parseInt(argument.split(" ")[2]);
+							indi.setAge(2017 - year);
 							indi.setBirthday(argument);
 							indiList.put(key, indi);
 						}else{
@@ -81,6 +82,8 @@ public class ParseGEDCOMFile {
 						tag = splitParts[1];
 						if(level == 2 && tag.equals("DATE")){
 							argument = splitParts[2];
+							int year = Integer.parseInt(argument.split(" ")[2]);
+							indi.setAge(year - 2017 + indi.getAge());
 							indi.setDeath(argument);
 							indiList.put(key, indi);
 						}else{
@@ -141,8 +144,10 @@ public class ParseGEDCOMFile {
 						tag = splitParts[1];
 						if(level == 2 && tag.equals("DATE")){
 							argument = splitParts[2];
+							if(marriageBeforeDivorce(fam, argument)){
 							fam.setDivorced(argument);
 							famList.put(key, fam);
+							}
 						}else{
 							break;
 						}
@@ -154,30 +159,36 @@ public class ParseGEDCOMFile {
 		}catch (IOException e) {
 			e.printStackTrace();
 		}
-		
-		System.out.println("INDI");
-		System.out.println("-------------------------------------");
+
+		System.out.println("Individuals");
+		System.out.println("  ID          Name         Gender    Birthday    Age   Alive     Death       Child     Spouse ");
+		System.out.println("====== ================== ======== ============ ===== ======= ============ ========= =========");
 		for(int i = 0;i< 5000;i++){
 			if(indiList.containsKey(i)){
 				Individual indi = indiList.get(i);
-				System.out.println(indi);
+				System.out.printf("  %-4s   %-16s    %-5s  %-11s  %-4s  %-6s  %-11s    %-6s    %-6s\n",
+						indi.getId(), indi.getName(), indi.getGender(), indi.getBirthday(),indi.getAge(), indi.getDeath() == "N/A" ? "True" : "False", indi.getDeath() == "N/A" ? "   N/A     " : indi.getDeath(), indi.getFamcId(), indi.getFamsId());
 			}
 		}
-		System.out.println("-------------------------------------");
-		System.out.println("FAM");
-		System.out.println("-------------------------------------");
+		System.out.println("====== ================== ======== ============ ===== ======= ============ ========= =========\n");
+
+		System.out.println("Families");
+		System.out.println("  ID     Married      Divorced    Husband ID     Husband Name     Wife ID      Wife Name          Children    ");
+		System.out.println("====== ============ ============ ============ ================== ========= ================== ================");
 		for(int i = 0;i< 1000;i++){
 			if(famList.containsKey(i)){
 				Family fam = famList.get(i);
-				System.out.println(fam);
+				System.out.printf("  %-4s  %-11s  %-11s     %-8s  %-17s    %-6s   %-16s    %-13s\n",
+						fam.getId(), fam.getMarried(), fam.getDivorced() == "N/A" ? "   N/A    " : fam.getDivorced(), fam.getHusId(), fam.getHusName(), fam.getWifId(), fam.getWifName(), fam.getChildren());
 			}
 		}
+		System.out.println("====== ============ ============ ============ ================== ========= ================== ================");
 	}
-	
+
 	public static boolean isValid(Integer _level, String _tag) {
-		int level = _level; 
+		int level = _level;
 		String tempTag = _tag;
-		
+
 		if (level == 0) {
 			List<String> list = Arrays.asList(level0Tag);
 			if (list.contains(tempTag)) {
@@ -204,240 +215,28 @@ public class ParseGEDCOMFile {
 		}
 		return false;
 	}
-	
-}
-	 class Individual{
-			private String id;
-			private String name;
-			private int age;
-			private String gender;
-			private String birthday;
-			private boolean Alive;
-			private String death = "N/A";
-			private String famcId = "N/A";
-			private String famsId = "N/A";
-			public Individual(){
-				setAlive(true);
-			}
-			public Individual(String id,String name,String gender,String birthday){
-				setId(id);
-				setName(name);
-				setGender(gender);
-				setBirthday(transDate(birthday));
-				setAlive(true);			
-			}
-			public String getId() {
-				return id;
-			}
-			public void setId(String id) {
-				this.id = id;
-			}
-			public String getName() {
-				return name;
-			}
-			public void setName(String name) {
-				this.name = name;
-			}
-			public int getAge() {
-				return age;
-			}
-			public void setAge(int age) {
-				this.age = age;
-			}
-			public String getGender() {
-				return gender;
-			}
-			public void setGender(String gender) {
-				this.gender = gender;
-			}
-			public String getBirthday() {
-				return birthday;
-			}
-			public void setBirthday(String birthday) {
-				this.birthday = transDate(birthday);
-			}
-			public boolean isAlive() {
-				return Alive;
-			}
-			public void setAlive(boolean alive) {
-				Alive = alive;
-			}
-			public String getDeath() {
-				return death;
-			}
-			public void setDeath(String death) {
-				this.death = transDate(death);
-			}
-			public String getFamcId() {
-				return famcId;
-			}
-			public void setFamcId(String famcId) {
-				this.famcId = famcId;
-			}
-			public String getFamsId() {
-				return famsId;
-			}
-			public void setFamsId(String famsId) {
-				this.famsId = famsId;
-			}
-			public String transDate(String date){
-				String split[] = date.split(" ");
-				String day = split[0];
-				int temp = Integer.parseInt(day);
-				if(temp<10){
-					day="0"+day;
+	public static boolean marriageBeforeDivorce(Family fam,String date){
+		date = fam.transDate(date);
+		String[] temp = date.split("-");
+		int year = Integer.parseInt(temp[0]);
+		int month = Integer.parseInt(temp[1]);
+		int day = Integer.parseInt(temp[2]);
+		if(year<fam.getmYear()){
+			return false;
+		}else if(year>fam.getmYear()){
+			return true;
+		}else{
+			if(month<fam.getmMonth()){
+				return false;
+			}else if(month>fam.getdMonth()){
+				return true;
+			}else{
+				if(day<=fam.getmDay()){
+					return false; 
+				}else{
+					return true;
 				}
-				String month=null;
-				if(split[1].equals("JAN")){
-					month="01";
-				}else if(split[1].equals("FEB")){
-					month="02";
-				}else if(split[1].equals("MAR")){
-					month="03";
-				}else if(split[1].equals("APR")){
-					month="04";
-				}else if(split[1].equals("MAY")){
-					month="05";
-				}else if(split[1].equals("JUN")){
-					month="06";
-				}else if(split[1].equals("JUL")){
-					month="07";
-				}else if(split[1].equals("AUG")){
-					month="08";
-				}else if(split[1].equals("SEP")){
-					month="09";
-				}else if(split[1].equals("OCT")){
-					month="10";
-				}else if(split[1].equals("NOV")){
-					month="11";
-				}else if(split[1].equals("DEC")){
-					month="12";
-				}
-				String year = split[2];
-				String d = year+"-"+month+"-"+day;	
-				return d;
 			}
-			@Override
-			public String toString() {
-				return "Individual [id=" + id + ", name=" + name + ", age="
-						+ age + ", gender=" + gender + ", birthday=" + birthday
-						+ ", Alive=" + Alive + ", death=" + death + ", famcId="
-						+ famcId + ", famsId=" + famsId + "]";
-			}
-			
-				
 		}
-		 class Family{
-			 private String id;
-			 private String married;
-			 private String divorced = "N/A";
-			 private String husId;
-			 private String wifId;
-			 private String husName;
-			 private String wifName;
-			 private ArrayList<String> children;
-			public Family(){
-				setChildren(new ArrayList<String>());
-			}
-			public Family(String id,String married,String husId,String wifId){
-				setId(id);
-				setMarried(married);
-				setHusId(husId);
-				setWifId(wifId);
-			}
-			public String getId() {
-				return id;
-			}
-			public void setId(String id) {
-				this.id = id;
-			}
-			public String getMarried() {
-				return married;
-			}
-			public void setMarried(String married) {
-				this.married = transDate(married);
-			}
-			public String getDivorced() {
-				return divorced;
-			}
-			public void setDivorced(String divorced) {
-				this.divorced = transDate(divorced);
-			}
-			public String getHusId() {
-				return husId;
-			}
-			public void setHusId(String husId) {
-				this.husId = husId;
-			}
-			public String getWifId() {
-				return wifId;
-			}
-			public void setWifId(String wifId) {
-				this.wifId = wifId;
-			}
-			public String getHusName() {
-				return husName;
-			}
-			public void setHusName(String husName) {
-				this.husName = husName;
-			}
-			public String getWifName() {
-				return wifName;
-			}
-			public void setWifName(String wifName) {
-				this.wifName = wifName;
-			}
-			
-			public ArrayList<String> getChildren() {
-				return children;
-			}
-			public void setChildren(ArrayList<String> children) {
-				this.children = children;
-			}
-			public String transDate(String date){
-				String split[] = date.split(" ");
-				String day = split[0];
-				int temp = Integer.parseInt(day);
-				if(temp<10){
-					day="0"+day;
-				}
-				String month=null;
-				if(split[1].equals("JAN")){
-					month="01";
-				}else if(split[1].equals("FEB")){
-					month="02";
-				}else if(split[1].equals("MAR")){
-					month="03";
-				}else if(split[1].equals("APR")){
-					month="04";
-				}else if(split[1].equals("MAY")){
-					month="05";
-				}else if(split[1].equals("JUN")){
-					month="06";
-				}else if(split[1].equals("JUL")){
-					month="07";
-				}else if(split[1].equals("AUG")){
-					month="08";
-				}else if(split[1].equals("SEP")){
-					month="09";
-				}else if(split[1].equals("OCT")){
-					month="10";
-				}else if(split[1].equals("NOV")){
-					month="11";
-				}else if(split[1].equals("DEC")){
-					month="12";
-				}
-				String year = split[2];
-				String d = year+"-"+month+"-"+day;	
-				return d;
-			}
-			@Override
-			public String toString() {
-				return "Family [id=" + id + ", married=" + married
-						+ ", divorced=" + divorced + ", husId=" + husId
-						+ ", wifId=" + wifId + ", husName=" + husName
-						+ ", wifName=" + wifName + ", children=" + children
-						+ "]";
-			}
-			
-		 }
+	}
+}
